requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
discord.py==1.7.3
requests==2.26.0
python-dotenv